+++ SBansheeEngine/Source/BsScriptManagedResource.cpp
@@ -1,5 +1,6 @@
 #include "BsScriptManagedResource.h"
 #include "BsScriptResourceManager.h"
+#include "BsScriptAssemblyManager.h"
 #include "BsScriptMeta.h"
 #include "BsMonoField.h"
 #include "BsMonoClass.h"
@@ -13,7 +14,11 @@ namespace BansheeEngine
 	ScriptManagedResource::ScriptManagedResource(MonoObject* instance, const HManagedResource& resource)
 		:ScriptObject(instance), mResource(resource)
 	{
+		assert(instance != nullptr);
 
+		::MonoClass* monoClass = mono_object_get_class(instance);
+		mNamespace = mono_class_get_namespace(monoClass);
+		mType = mono_class_get_name(monoClass);
 	}
 
 	void ScriptManagedResource::initRuntimeData()
@@ -28,18 +33,54 @@ namespace BansheeEngine
 		ScriptResourceManager::instance().createManagedResource(instance, resource);
 	}
 
+	MonoObject* ScriptManagedResource::_createManagedInstance(bool construct)
+	{
+		ManagedSerializableObjectInfoPtr currentObjInfo = nullptr;
+
+		// See if this type even still exists
+		if (!ScriptAssemblyManager::instance().getSerializableObjectInfo(mNamespace, mType, currentObjInfo))
+			return nullptr;
+
+		return currentObjInfo->mMonoClass->createInstance(construct);
+	}
+
+	ScriptObjectBackup ScriptManagedResource::beginRefresh()
+	{
+		ScriptResourceBase::beginRefresh();
+
+		ScriptObjectBackup backupData;
+		backupData.data = mResource->backup(true);
+
+		return backupData;
+	}
+
+	void ScriptManagedResource::endRefresh(const ScriptObjectBackup& backupData)
+	{
+		ResourceBackupData resourceBackup = any_cast<ResourceBackupData>(backupData.data);
+		mResource->restore(mManagedInstance, resourceBackup);
+
+		// If we could not find resource type after refresh, treat it as if it was destroyed
+		if (mManagedInstance == nullptr)
+			_onManagedInstanceDeleted();
+
+		ScriptResourceBase::endRefresh(backupData);
+	}
+
 	void ScriptManagedResource::_onManagedInstanceDeleted()
 	{
 		mManagedInstance = nullptr;
 		
-		// The only way this method should be reachable is when Resource::unload is called, which means the resource
-		// has had to been already freed. Even if all managed instances are released ManagedResource itself holds the last
-		// instance which is only freed on unload().
-		// Note: During domain unload this could get called even if not all instances are released, but ManagedResourceManager
-		// should make sure all instances are unloaded before that happens.
-		assert(mResource == nullptr || !mResource.isLoaded()); 
-		
-		ScriptResourceManager::instance().destroyScriptResource(this);
+		if (!mRefreshInProgress)
+		{
+			// The only way this method should be reachable is when Resource::unload is called, which means the resource
+			// has had to been already freed. Even if all managed instances are released ManagedResource itself holds the last
+			// instance which is only freed on unload().
+			// Note: During domain unload this could get called even if not all instances are released, but ManagedResourceManager
+			// should make sure all instances are unloaded before that happens.
+			assert(mResource == nullptr || !mResource.isLoaded());
+
+			ScriptResourceManager::instance().destroyScriptResource(this);
+		}
 	}
 
 	void ScriptManagedResource::setNativeHandle(const HResource& resource)
